refactor: simplify detail pages and add copyable-field component
