+++ platform/flowglad-next/src/app/features/[id]/InnerFeatureDetailsPage.tsx
@@ -1,11 +1,12 @@
 'use client'
 
-import { Check, Copy, DollarSign, X } from 'lucide-react'
+import { Check, DollarSign, X } from 'lucide-react'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
 import EditFeatureModal from '@/components/forms/EditFeatureModal'
 import InnerPageContainerNew from '@/components/InnerPageContainerNew'
+import { CopyableField } from '@/components/ui/copyable-field'
 import { PageHeaderNew } from '@/components/ui/page-header-new'
 import { Feature } from '@/db/schema/features'
 import { PricingModel } from '@/db/schema/pricingModels'
@@ -17,56 +18,6 @@ interface InnerFeatureDetailsPageProps {
   feature: Feature.ClientRecord
   pricingModel: PricingModel.ClientRecord | null
   usageMeter: UsageMeter.ClientRecord | null
-}
-
-/**
- * Copyable field component for displaying values with a copy button.
- * Based on Figma design - copy icon is always visible.
- */
-function CopyableField({
-  value,
-  label,
-}: {
-  value: string
-  label: string
-}) {
-  const [copied, setCopied] = useState(false)
-
-  const handleCopy = async () => {
-    try {
-      await navigator.clipboard.writeText(value)
-      setCopied(true)
-      setTimeout(() => setCopied(false), 2000)
-    } catch (error) {
-      console.error('Failed to copy:', error)
-    }
-  }
-
-  return (
-    <div
-      className="inline-flex items-center gap-1 cursor-pointer group"
-      onClick={handleCopy}
-      role="button"
-      tabIndex={0}
-      onKeyDown={(e) => {
-        if (e.key === 'Enter' || e.key === ' ') {
-          e.preventDefault()
-          handleCopy()
-        }
-      }}
-      aria-label={`Copy ${label}`}
-      title={`Click to copy ${label}`}
-    >
-      {copied ? (
-        <Check className="h-4 w-4 text-[hsl(var(--jade-muted-foreground))] flex-shrink-0" />
-      ) : (
-        <Copy className="h-4 w-4 text-foreground flex-shrink-0" />
-      )}
-      <span className="font-sans font-normal text-base leading-6 text-foreground group-hover:underline transition-colors">
-        {value}
-      </span>
-    </div>
-  )
 }
 
 /**
