+++ platform/flowglad-next/src/app/pricing-models/[id]/InnerPricingModelDetailsPage.tsx
@@ -1,14 +1,7 @@
 'use client'
-import {
-  Check,
-  Copy,
-  Download,
-  Pencil,
-  Sparkles,
-  Star,
-} from 'lucide-react'
+import { Check, Copy, Download, Pencil, Sparkles } from 'lucide-react'
 import { useRouter } from 'next/navigation'
-import { useEffect, useRef, useState } from 'react'
+import { useState } from 'react'
 import { toast } from 'sonner'
 import { trpc } from '@/app/_trpc/client'
 import { CustomersDataTable } from '@/app/customers/data-table'
@@ -29,94 +22,14 @@ import { MoreIcon } from '@/components/icons/MoreIcon'
 import PopoverMenu, {
   type PopoverMenuItem,
 } from '@/components/PopoverMenu'
+import { CopyableField } from '@/components/ui/copyable-field'
 import { PageHeaderNew } from '@/components/ui/page-header-new'
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover'
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/ui/tooltip'
 import type { PricingModel } from '@/db/schema/pricingModels'
-
-/**
- * Copyable field component for displaying values with a copy button.
- */
-function CopyableField({
-  value,
-  label,
-  displayText,
-}: {
-  value: string
-  label: string
-  displayText?: string
-}) {
-  const [copied, setCopied] = useState(false)
-  const timeoutRef = useRef<ReturnType<typeof setTimeout> | null>(
-    null
-  )
-
-  useEffect(() => {
-    return () => {
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current)
-      }
-    }
-  }, [])
-
-  const handleCopy = async () => {
-    try {
-      await navigator.clipboard.writeText(value)
-      setCopied(true)
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current)
-      }
-      timeoutRef.current = setTimeout(() => setCopied(false), 2000)
-    } catch (error) {
-      console.error('Failed to copy:', error)
-    }
-  }
-
-  return (
-    <TooltipProvider delayDuration={300}>
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <div
-            className="inline-flex items-center gap-1 cursor-pointer group"
-            onClick={handleCopy}
-            role="button"
-            tabIndex={0}
-            onKeyDown={(e) => {
-              if (e.key === 'Enter' || e.key === ' ') {
-                e.preventDefault()
-                handleCopy()
-              }
-            }}
-            aria-label={`Copy ${label}`}
-          >
-            {copied ? (
-              <Check className="h-3.5 w-3.5 text-[hsl(var(--jade-muted-foreground))] flex-shrink-0" />
-            ) : (
-              <Copy className="h-3.5 w-3.5 text-muted-foreground flex-shrink-0" />
-            )}
-            <span className="font-sans font-medium text-sm leading-5 text-muted-foreground group-hover:underline transition-colors">
-              {copied && displayText
-                ? displayText.replace(/^Copy/, 'Copied')
-                : (displayText ?? value)}
-            </span>
-          </div>
-        </TooltipTrigger>
-        <TooltipContent side="bottom">
-          <p className="font-sans">{value}</p>
-        </TooltipContent>
-      </Tooltip>
-    </TooltipProvider>
-  )
-}
 
 export type InnerPricingModelDetailsPageProps = {
   pricingModel: PricingModel.ClientRecord
@@ -244,7 +157,7 @@ function InnerPricingModelDetailsPage({
               setIsMoreMenuOpen(false)
               setIsSetDefaultOpen(true)
             },
-            icon: <Star className="h-4 w-4" />,
+            icon: <Check className="h-4 w-4" />,
           },
         ]
       : []),
@@ -311,8 +224,8 @@ function InnerPricingModelDetailsPage({
                     }}
                     aria-label="More options"
                   >
-                    <MoreIcon className="h-3.5 w-3.5 text-muted-foreground flex-shrink-0" />
-                    <span className="font-sans font-medium text-sm leading-5 text-muted-foreground group-hover:underline transition-colors">
+                    <MoreIcon className="h-3.5 w-3.5 text-muted-foreground group-hover:text-foreground flex-shrink-0 transition-colors" />
+                    <span className="font-sans font-medium text-sm leading-5 text-muted-foreground group-hover:underline group-hover:text-foreground transition-colors">
                       More options
                     </span>
                   </div>
@@ -329,7 +242,6 @@ function InnerPricingModelDetailsPage({
           title="Products"
           defaultExpanded={true}
           contentPadding={false}
-          borderTop
         >
           <ProductsDataTable
             filters={getProductFilterForTab(activeProductFilter)}
